--- Assets/scripts/valo_machine.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9b6a493a0c546684184683b0d01457ce
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
