+++ Assets/soundeffects/Music.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9619d7e6c20d1f44aa75c86b0d9f73e0
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
