+++ Assets/soundeffects/Tietoiskut.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f5b8aec062d210c4094f7639a8681262
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
