+++ Assets/soundeffects/Tietoiskut/tietoiskuTV.mp3.meta
@@ -1,0 +1,22 @@
+fileFormatVersion: 2
+guid: 9f8396e722a51dd4ab2cc98182f7ae31
+AudioImporter:
+  externalObjects: {}
+  serializedVersion: 6
+  defaultSettings:
+    loadType: 0
+    sampleRateSetting: 0
+    sampleRateOverride: 44100
+    compressionFormat: 1
+    quality: 1
+    conversionMode: 0
+  platformSettingOverrides: {}
+  forceToMono: 0
+  normalize: 1
+  preloadAudioData: 1
+  loadInBackground: 0
+  ambisonic: 0
+  3D: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
