+++ Assets/soundeffects/laitteet yms.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7248ccf4f8a61294aa089c4cf96e8c1a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
