+++ Assets/soundeffects/laitteet yms/Click2-Sebastian-759472264.mp3.meta
@@ -1,0 +1,22 @@
+fileFormatVersion: 2
+guid: 70e628aad14cb6743abfa5f1a4aca8da
+AudioImporter:
+  externalObjects: {}
+  serializedVersion: 6
+  defaultSettings:
+    loadType: 0
+    sampleRateSetting: 0
+    sampleRateOverride: 44100
+    compressionFormat: 1
+    quality: 1
+    conversionMode: 0
+  platformSettingOverrides: {}
+  forceToMono: 0
+  normalize: 1
+  preloadAudioData: 1
+  loadInBackground: 0
+  ambisonic: 0
+  3D: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
