+++ Assets/soundeffects/laitteet yms/Clothes Dryer-SoundBible.com-962232937.mp3.meta
@@ -1,0 +1,22 @@
+fileFormatVersion: 2
+guid: 17a1649d3cee8ca45984985f4ac76351
+AudioImporter:
+  externalObjects: {}
+  serializedVersion: 6
+  defaultSettings:
+    loadType: 0
+    sampleRateSetting: 0
+    sampleRateOverride: 44100
+    compressionFormat: 1
+    quality: 1
+    conversionMode: 0
+  platformSettingOverrides: {}
+  forceToMono: 0
+  normalize: 1
+  preloadAudioData: 1
+  loadInBackground: 0
+  ambisonic: 0
+  3D: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
