+++ Assets/soundeffects/laitteet yms/Dishwasher-SoundBible.com-1917190099.mp3.meta
@@ -1,0 +1,22 @@
+fileFormatVersion: 2
+guid: 0598b8ee980623b4ebf41e440c963140
+AudioImporter:
+  externalObjects: {}
+  serializedVersion: 6
+  defaultSettings:
+    loadType: 0
+    sampleRateSetting: 0
+    sampleRateOverride: 44100
+    compressionFormat: 1
+    quality: 1
+    conversionMode: 0
+  platformSettingOverrides: {}
+  forceToMono: 0
+  normalize: 1
+  preloadAudioData: 1
+  loadInBackground: 0
+  ambisonic: 0
+  3D: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
